* Reorganize to acommodate TF2.x parser.
